Implement more tests for character movement/collisions
NOTE Non of the code in my sandbox is meant to be used for the
real game. In fact I would rather just remake them from scratch or
possibly delete them later. Reason I am doing this is to share my
testing environment for movement/collisions so that other team
members can see the different implications of different
combinations of RB/CC/Colliders/Triggers, since the Unity documentation
is lacking when it comes to CC + collisions/triggers.