Fix bug where the player wouldnt animate after possess
Not exactly sure why this bug was occuring but I did find out
that the player had two models (old one and new one) in it's
hierarchy immediately after possession. Even after the destroy
methods were called on the model, it was still there (I assume
the actual destroy was occuring 1 or 2 frames later). This caused
the previous model's animator to be referenced into ani, which would
only get destroyed and thus no animation would play on the player.